Consejero/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
py-cpuinfo==8.0.0
psutil==5.9.0
GPUtil==1.4.0
gunicorn==20.1.0